Add Dockerfiles for both external libs and glue build Changes to be committed: 	modified:   .gitignore 	new file:   Dockerfile 	new file:   README_DOCKER 	modified:   conf/config.mk 	new file:   external/Dockerfile 	new file:   gmp.properties 	modified:   src/examples/gmp.properties
Summary:

Added two Dockerfiles for both the external libs needed and glue build up with gcc-4.9
Also the config files with paths (config.mk) and with the location for the GMP (gmp.properties) were changed to work with the docker-compose setup. If other file and volumes are used the files need modifications and images need to be rebuild.

- stroncod